+++ spec/support/ssl_helper.rb
@@ -33,8 +33,7 @@ module SSLHelper
     end
 
     def generate_ecdsa_key(curve)
-      k = OpenSSL::PKey::EC.new(curve)
-      k.generate_key
+      k = OpenSSL::PKey::EC.generate(curve)
       Acme::Client::CertificateRequest::ECKeyPatch.new(k)
     end
 
@@ -104,16 +103,21 @@ module SSLHelper
   # priv - An OpenSSL::PKey::EC or OpenSSL::PKey::RSA instance.
   #
   # Returns a String.
-  def public_key_to_der(priv)
-    case priv
+  def public_key_to_pem(private_key)
+    case private_key
     when OpenSSL::PKey::EC
-      dup = OpenSSL::PKey::EC.new(priv.to_der)
-      dup.private_key = nil
-      dup.to_der
+      # TODO: Ruby 2.7 shenanigans
+      if OpenSSL::PKey::EC.method_defined?(:to_pem)
+        private_key.to_pem
+      else
+        dup = OpenSSL::PKey::EC.new(private_key.to_der)
+        dup.private_key = nil
+        dup.to_pem
+      end
     when OpenSSL::PKey::RSA
-      priv.public_key.to_der
+      private_key.public_key.to_pem
     else
-      raise ArgumentError, 'priv must be EC or RSA'
+      raise ArgumentError, 'private_key must be EC or RSA'
     end
   end
 end